Move endpoints from prefix /api/master/block to /api/master
