+++ core/server/src/main/java/alluxio/master/AlluxioMasterRestServiceHandler.java
@@ -14,13 +14,18 @@ package alluxio.master;
 import alluxio.Configuration;
 import alluxio.Constants;
 import alluxio.Version;
+import alluxio.master.block.BlockMaster;
+import alluxio.underfs.UnderFileSystem;
 import alluxio.util.CommonUtils;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import com.codahale.metrics.Counter;
 import com.codahale.metrics.Gauge;
 import com.codahale.metrics.MetricRegistry;
 import com.qmino.miredot.annotations.ReturnType;
 
+import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Set;
@@ -40,17 +45,32 @@ import javax.ws.rs.core.Response;
 @Produces(MediaType.APPLICATION_JSON)
 // TODO(cc): Investigate auto-generation of REST API documentation.
 public final class AlluxioMasterRestServiceHandler {
+  private static final Logger LOG = LoggerFactory.getLogger(Constants.LOGGER_TYPE);
+
   public static final String SERVICE_PREFIX = "master";
+  public static final String GET_RPC_ADDRESS = "rpc_address";
   public static final String GET_CONFIGURATION = "configuration";
   public static final String GET_DEBUG = "debug";
-  public static final String GET_ADDRESS = "address";
+  public static final String GET_CAPACITY_BYTES = "capacity_bytes";
+  public static final String GET_USED_BYTES = "used_bytes";
+  public static final String GET_FREE_BYTES = "free_bytes";
+  public static final String GET_CAPACITY_BYTES_ON_TIERS = "capacity_bytes_on_tiers";
+  public static final String GET_USED_BYTES_ON_TIERS = "used_bytes_on_tiers";
+  public static final String GET_DISK_CAPACITY_BYTES = "disk_capacity_bytes";
+  public static final String GET_DISK_FREE_BYTES = "disk_free_bytes";
+  public static final String GET_DISK_USED_BYTES = "disk_used_bytes";
   public static final String GET_METRICS = "metrics";
   public static final String GET_START_TIME_MS = "start_time_ms";
   public static final String GET_UPTIME_MS = "uptime_ms";
   public static final String GET_VERSION = "version";
+  public static final String GET_WORKER_COUNT = "worker_count";
+  public static final String GET_WORKER_INFO_LIST = "worker_info_list";
 
   private final AlluxioMaster mMaster = AlluxioMaster.get();
+  private final BlockMaster mBlockMaster = mMaster.getBlockMaster();
   private final Configuration mMasterConf = MasterContext.getConf();
+  private final String mUfsRoot = mMasterConf.get(Constants.UNDERFS_ADDRESS);
+  private final UnderFileSystem mUfs = UnderFileSystem.get(mUfsRoot, mMasterConf);
 
   /**
    * @summary get the configuration map
@@ -81,14 +101,14 @@ public final class AlluxioMasterRestServiceHandler {
   }
 
   /**
-   * @summary get the master address
+   * @summary get the master rpc address
    * @return the response object
    */
   @GET
-  @Path(GET_ADDRESS)
+  @Path(GET_RPC_ADDRESS)
   @ReturnType("java.lang.String")
-  public Response getAddress() {
-    return Response.ok(mMasterConf.get(Constants.MASTER_ADDRESS)).build();
+  public Response getRpcAddress() {
+    return Response.ok(mMaster.getMasterAddress().toString()).build();
   }
 
   /**
@@ -115,7 +135,7 @@ public final class AlluxioMasterRestServiceHandler {
     for (Map.Entry<String, Counter> counter : counters.entrySet()) {
       metrics.put(counter.getKey(), counter.getValue().getCount());
     }
-    metrics.put(filesPinnedProperty, (Long) filesPinned.getValue());
+    metrics.put(filesPinnedProperty, ((Integer) filesPinned.getValue()).longValue());
 
     return Response.ok(metrics).build();
   }
@@ -151,5 +171,130 @@ public final class AlluxioMasterRestServiceHandler {
   @ReturnType("java.lang.String")
   public Response getVersion() {
     return Response.ok(Version.VERSION).build();
+  }
+
+  /**
+   * @summary get the total capacity of all workers in bytes
+   * @return the response object
+   */
+  @GET
+  @Path(GET_CAPACITY_BYTES)
+  @ReturnType("java.lang.Long")
+  public Response getCapacityBytes() {
+    return Response.ok(mBlockMaster.getCapacityBytes()).build();
+  }
+
+  /**
+   * @summary get the total disk capacity in bytes, a negative value means the capacity is unknown.
+   * @return the response object
+   */
+  @GET
+  @Path(GET_DISK_CAPACITY_BYTES)
+  @ReturnType("java.lang.Long")
+  public Response getDiskCapacityBytes() {
+    try {
+      return Response.ok(mUfs.getSpace(mUfsRoot, UnderFileSystem.SpaceType.SPACE_TOTAL)).build();
+    } catch (IOException e) {
+      LOG.warn(e.getMessage());
+      return Response.serverError().entity(e.getMessage()).build();
+    }
+  }
+
+  /**
+   * @summary get the free disk capacity in bytes, a negative value means the capacity is unknown.
+   * @return the response object
+   */
+  @GET
+  @Path(GET_DISK_FREE_BYTES)
+  @ReturnType("java.lang.Long")
+  public Response getFreeDiskCapacityBytes() {
+    try {
+      return Response.ok(mUfs.getSpace(mUfsRoot, UnderFileSystem.SpaceType.SPACE_FREE)).build();
+    } catch (IOException e) {
+      LOG.warn(e.getMessage());
+      return Response.serverError().entity(e.getMessage()).build();
+    }
+  }
+
+  /**
+   * @summary get the used disk capacity, a negative value means the capacity is unknown.
+   * @return the response object
+   */
+  @GET
+  @Path(GET_DISK_USED_BYTES)
+  @ReturnType("java.lang.Long")
+  public Response getUsedDiskCapacityBytes() {
+    try {
+      return Response.ok(mUfs.getSpace(mUfsRoot, UnderFileSystem.SpaceType.SPACE_USED)).build();
+    } catch (IOException e) {
+      LOG.warn(e.getMessage());
+      return Response.serverError().entity(e.getMessage()).build();
+    }
+  }
+
+  /**
+   * @summary get the free capacity
+   * @return the response object
+   */
+  @GET
+  @Path(GET_FREE_BYTES)
+  @ReturnType("java.lang.Long")
+  public Response getFreeCapacityBytes() {
+    return Response.ok(mBlockMaster.getCapacityBytes() - mBlockMaster.getUsedBytes()).build();
+  }
+
+  /**
+   * @summary get the mapping from tier alias to total capacity of the tier in bytes
+   * @return the response object
+   */
+  @GET
+  @Path(GET_CAPACITY_BYTES_ON_TIERS)
+  @ReturnType("java.util.Map<String, Long>")
+  public Response getCapacityBytesOnTiers() {
+    return Response.ok(mBlockMaster.getTotalBytesOnTiers()).build();
+  }
+
+  /**
+   * @summary get the used capacity
+   * @return the response object
+   */
+  @GET
+  @Path(GET_USED_BYTES)
+  @ReturnType("java.lang.Long")
+  public Response getUsedBytes() {
+    return Response.ok(mBlockMaster.getUsedBytes()).build();
+  }
+
+  /**
+   * @summary get the mapping from tier alias to the used bytes of the tier
+   * @return the response object
+   */
+  @GET
+  @Path(GET_USED_BYTES_ON_TIERS)
+  @ReturnType("java.util.Map<String, Long>")
+  public Response getUsedBytesOnTiers() {
+    return Response.ok(mBlockMaster.getUsedBytesOnTiers()).build();
+  }
+
+  /**
+   * @summary get the count of workers
+   * @return the response object
+   */
+  @GET
+  @Path(GET_WORKER_COUNT)
+  @ReturnType("java.lang.Integer")
+  public Response getWorkerCount() {
+    return Response.ok(mBlockMaster.getWorkerCount()).build();
+  }
+
+  /**
+   * @summary get the list of worker descriptors
+   * @return the response object
+   */
+  @GET
+  @Path(GET_WORKER_INFO_LIST)
+  @ReturnType("java.util.List<alluxio.wire.WorkerInfo>")
+  public Response getWorkerInfoList() {
+    return Response.ok(mBlockMaster.getWorkerInfoList()).build();
   }
 }
