+++ tests/src/test/java/alluxio/master/AlluxioMasterRestApiTest.java
@@ -9,15 +9,15 @@
  * See the NOTICE file distributed with this work for information regarding copyright ownership.
  */
 
-package alluxio.master.block;
+package alluxio.master;
 
 import alluxio.LocalAlluxioClusterResource;
-import alluxio.Constants;
-import alluxio.master.AlluxioMaster;
+import alluxio.master.block.BlockMaster;
 import alluxio.rest.TestCaseFactory;
-import alluxio.wire.BlockInfo;
-import alluxio.wire.BlockInfoTest;
+import alluxio.wire.WorkerInfo;
+import alluxio.wire.WorkerInfoTest;
 
+import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 import org.junit.BeforeClass;
 import org.junit.Ignore;
@@ -30,15 +30,17 @@ import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.modules.junit4.PowerMockRunner;
 import org.powermock.reflect.Whitebox;
 
+import java.util.List;
 import java.util.Map;
+import java.util.Random;
 
 /**
- * Test cases for {@link BlockMasterClientRestServiceHandler}.
+ * Test cases for {@link AlluxioMasterRestServiceHandler}.
  */
 @RunWith(PowerMockRunner.class)
 @PrepareForTest(BlockMaster.class)
 @Ignore
-public class BlockMasterClientRestApiTest {
+public class AlluxioMasterRestApiTest {
   private static final Map<String, String> NO_PARAMS = Maps.newHashMap();
   private static BlockMaster sBlockMaster;
 
@@ -54,35 +56,49 @@ public class BlockMasterClientRestApiTest {
   }
 
   private String getEndpoint(String suffix) {
-    return BlockMasterClientRestServiceHandler.SERVICE_PREFIX + "/" + suffix;
+    return AlluxioMasterRestServiceHandler.SERVICE_PREFIX + "/" + suffix;
   }
 
   @Test
-  public void serviceNameTest() throws Exception {
+  public void getCapacityBytesTest() throws Exception {
+    Random random = new Random();
+    long capacityBytes = random.nextLong();
+    Mockito.doReturn(capacityBytes).when(sBlockMaster).getCapacityBytes();
+
     TestCaseFactory
-        .newMasterTestCase(getEndpoint(BlockMasterClientRestServiceHandler.SERVICE_NAME), NO_PARAMS,
-            "GET", Constants.BLOCK_MASTER_CLIENT_SERVICE_NAME, mResource).run();
+        .newMasterTestCase(getEndpoint(AlluxioMasterRestServiceHandler.GET_CAPACITY_BYTES),
+            NO_PARAMS, "GET", capacityBytes, mResource).run();
+
+    Mockito.verify(sBlockMaster).getCapacityBytes();
   }
 
   @Test
-  public void serviceVersionTest() throws Exception {
+  public void getUsedBytesTest() throws Exception {
+    Random random = new Random();
+    long usedBytes = random.nextLong();
+    Mockito.doReturn(usedBytes).when(sBlockMaster).getUsedBytes();
+
     TestCaseFactory
-        .newMasterTestCase(getEndpoint(BlockMasterClientRestServiceHandler.SERVICE_VERSION),
-            NO_PARAMS, "GET", Constants.BLOCK_MASTER_CLIENT_SERVICE_VERSION, mResource).run();
+        .newMasterTestCase(getEndpoint(AlluxioMasterRestServiceHandler.GET_USED_BYTES),
+            NO_PARAMS, "GET", usedBytes, mResource).run();
+
+    Mockito.verify(sBlockMaster).getUsedBytes();
   }
 
   @Test
-  public void getBlockInfoTest() throws Exception {
-    Map<String, String> params = Maps.newHashMap();
-    params.put("blockId", "1");
-
-    BlockInfo blockInfo = BlockInfoTest.createRandom();
-    Mockito.doReturn(blockInfo).when(sBlockMaster).getBlockInfo(Mockito.anyLong());
+  public void getWorkerInfoListTest() throws Exception {
+    Random random = new Random();
+    List<WorkerInfo> workerInfos = Lists.newArrayList();
+    int numWorkerInfos = random.nextInt(10);
+    for (int i = 0; i < numWorkerInfos; i++) {
+      workerInfos.add(WorkerInfoTest.createRandom());
+    }
+    Mockito.doReturn(workerInfos).when(sBlockMaster).getWorkerInfoList();
 
     TestCaseFactory
-        .newMasterTestCase(getEndpoint(BlockMasterClientRestServiceHandler.GET_BLOCK_INFO), params,
-            "GET", blockInfo, mResource).run();
+        .newMasterTestCase(getEndpoint(AlluxioMasterRestServiceHandler.GET_WORKER_INFO_LIST),
+            NO_PARAMS, "GET", workerInfos, mResource).run();
 
-    Mockito.verify(sBlockMaster).getBlockInfo(Mockito.anyLong());
+    Mockito.verify(sBlockMaster).getWorkerInfoList();
   }
 }
